refactor: break circular dependency in event system
Extract EventBus class and singleton instance into src/utils/event-bus.js
so that terminal-events.js and workspace-events.js no longer import from
events.js. events.js remains the backward-compatible aggregator, re-exporting
the bus from event-bus.js.

Dependency graph after this change:
  event-bus.js (zero imports)
    <- terminal-events.js
    <- workspace-events.js
    <- events.js (aggregation only)

Closes #263
