+++ src/utils/terminal-events.js
@@ -8,7 +8,7 @@
  * @see events.js (backward-compat re-exports)
  */
 
-import { bus } from './events.js';
+import { bus } from './event-bus.js';
 
 // ── Event constants ─────────────────────────────────────────────────
